FE/Qt4: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@34523 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMouseHandler.cpp
@@ -117,7 +117,7 @@ void UIMouseHandler::prepareListener(ulong uIndex, UIMachineWindow *pMachineWind
 void UIMouseHandler::cleanupListener(ulong uIndex)
 {
     /* Check if we should release mouse first: */
-    if (uIndex == m_iMouseCaptureViewIndex)
+    if ((int)uIndex == m_iMouseCaptureViewIndex)
         releaseMouse();
 
     /* If that window still registered: */
